bsc: timeslot_fsm: Handle and ignore tear down of lchan during PDCH DEACT
lchan sends TS_EV_LCHAN_UNUSED to its parent (ts) during release time. It
was experimentally found that it can happen that an lchan can be terminated
while waiting for a PDCH DEACT (N)ACK response.

The fsm definition actually states that this event can be received in
state TS_ST_WAIT_PDCH_DEACT, but it was not handled before and as a result the
process aborted due to the defalt switch case.

Change-Id: If61493e7d5449bf2c2de9fd34cdf2410625e92ac