Instead of T23042, use T-2427001
T23042 is obviously a phantasy number, so far used for various lchan and RTP
stream related timeouts. It's not a good idea to keep this around.

Since the timeout is mostly responsible for waiting for MGW endpoint setup,
instead, use 2427001 as timer number: 2427 is the default MGCP port, and 001 is
our first custom timer for that.

To indicate a non-standard timer number, actually use negative T-2427001.
(osmo_tdef tests in libosmocore show that this works.)

Change-Id: Id00b75dd8c486f8a4951afa49a03a1066b463aef