lchan_fsm: do not include 12.2k in active set on HR channels
When "Config-NB-Code = 1" is set via the S-bits, then the resulting
multirate configuration IE will contain 12.2K. Since the generator
function is not aware if the lchan is activated for HR or FR it sets the
flag for 12.2k always.

We have to add a check here in order to set the 12.2k flag in the IE
back to 0 so that the active set contains a set of useable rates.

Change-Id: I40e7f568f4822040a2d1e78f22dbba0e49d0167e
Related: SYS#4470